+++ net.tourbook/src/net/tourbook/ui/views/tourCatalog/TourPropertyRefTourChanged.java
@@ -1,17 +1,17 @@
 /*******************************************************************************
- * Copyright (C) 2005, 2009  Wolfgang Schramm and Contributors
- *   
+ * Copyright (C) 2005, 2011  Wolfgang Schramm and Contributors
+ * 
  * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software 
+ * the terms of the GNU General Public License as published by the Free Software
  * Foundation version 2 of the License.
- *  
- * This program is distributed in the hope that it will be useful, but WITHOUT 
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS 
+ * 
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
  * 
- * You should have received a copy of the GNU General Public License along with 
+ * You should have received a copy of the GNU General Public License along with
  * this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110, USA    
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110, USA
  *******************************************************************************/
 package net.tourbook.ui.views.tourCatalog;
 
@@ -20,7 +20,9 @@ import net.tourbook.ui.tourChart.TourChart;
 public class TourPropertyRefTourChanged {
 
 	long		refId;
-	int			xMarkerValue;
+
+	float		xMarkerValue;
+
 	TourChart	refTourChart;
 
 	/**
@@ -31,9 +33,11 @@ public class TourPropertyRefTourChanged {
 	 * @param refTourXMarkerValue
 	 *            value difference in the reference tour
 	 */
-	public TourPropertyRefTourChanged(final TourChart refTourChart, final long refId, final int refTourXMarkerValue) {
+	public TourPropertyRefTourChanged(final TourChart refTourChart, final long refId, final float refTourXMarkerValue) {
+
 		this.refTourChart = refTourChart;
 		this.refId = refId;
+
 		this.xMarkerValue = refTourXMarkerValue;
 	}
